+++ .env
@@ -10,6 +10,6 @@ VITE_GITHUB_REDIRECT_URI=http://localhost:5173
 # В dev: http://localhost:3000 (mock-server)
 # В prod: https://your-domain.com (реальный бэкенд преподавателя)
 #VITE_API_URL=http://dancv.ddns.net
-VITE_API_URL=http://localhost:3000
+VITE_API_URL=http://83.246.202.150
 #VITE_API_URL=http://83.246.202.150
