+++ quiz-backend/package.json
@@ -19,8 +19,8 @@
     "zod": "^4.3.6"
   },
   "devDependencies": {
-    "@types/node": "^20.19.33",
-    "tsx": "^4.7.1",
+    "@types/node": "^20.19.37",
+    "tsx": "^4.21.0",
     "typescript": "^5.8.3",
     "vitest": "^4.0.18"
   }
